Harmonize RuntimeHintsRegistrar implementations
Closes gh-28801
