+++ spring-beans/src/main/resources/META-INF/spring/aot.factories
@@ -1,4 +1,5 @@
 org.springframework.beans.factory.aot.BeanFactoryInitializationAotProcessor=\
 org.springframework.beans.factory.aot.BeanRegistrationsAotProcessor
+
 org.springframework.aot.hint.RuntimeHintsRegistrar=\
-org.springframework.beans.factory.annotation.JakartaAnnotationsRuntimeHintsRegistrar
+org.springframework.beans.factory.annotation.JakartaAnnotationsRuntimeHints
