+++ spring-context-support/src/main/resources/META-INF/spring/aot.factories
@@ -1,2 +1,2 @@
 org.springframework.aot.hint.RuntimeHintsRegistrar= \
-org.springframework.mail.javamail.MimeTypeResourceHintsRegistrar
+org.springframework.mail.javamail.JavaMailMimeTypesRuntimeHints
