--- spring-messaging/src/main/java/org/springframework/messaging/simp/annotation/SimpAnnotationsRuntimeHintsRegistrar.java
@@ -1,39 +1,0 @@
-/*
- * Copyright 2002-2022 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.springframework.messaging.simp.annotation;
-
-import java.util.stream.Stream;
-
-import org.springframework.aot.hint.RuntimeHints;
-import org.springframework.aot.hint.RuntimeHintsRegistrar;
-import org.springframework.aot.hint.support.RuntimeHintsUtils;
-
-/**
- * {@link RuntimeHintsRegistrar} implementation that makes Simp annotations
- * available at runtime.
- *
- * @author Sebastien Deleuze
- * @since 6.0
- */
-public class SimpAnnotationsRuntimeHintsRegistrar implements RuntimeHintsRegistrar {
-
-	@Override
-	public void registerHints(RuntimeHints hints, ClassLoader classLoader) {
-		Stream.of(SendToUser.class, SubscribeMapping.class).forEach(
-				annotationType -> RuntimeHintsUtils.registerAnnotation(hints, annotationType));
-	}
-}
