+++ spring-websocket/src/main/resources/META-INF/spring/aot.factories
@@ -1,2 +1,2 @@
 org.springframework.aot.hint.RuntimeHintsRegistrar= \
-org.springframework.web.socket.server.support.HandshakeHandlerRuntimeHintsRegistrar
+org.springframework.web.socket.server.support.HandshakeHandlerRuntimeHints
